:truck: chore(api) move certification challenges service
